+++ Modules/Intent.Modules.ModuleBuilder/Templates/IModSpec/TemplateRegistrationRequiredEvent.cs
@@ -5,13 +5,15 @@ namespace Intent.Modules.ModuleBuilder.Templates.IModSpec
 {
     public class TemplateRegistrationRequiredEvent
     {
-        public TemplateRegistrationRequiredEvent(string modelId, string templateId, string templateType, string role, string location)
+        public TemplateRegistrationRequiredEvent(string modelId, string templateId, string templateType, string role, string location, string severity, string classification)
         {
             ModelId = modelId;
             TemplateId = templateId;
             TemplateType = templateType;
             Role = role;
             Location = location;
+            Severity = severity;
+            Classification = classification;
         }
 
         public TemplateRegistrationRequiredEvent(IModuleBuilderTemplate template)
@@ -23,6 +25,8 @@ namespace Intent.Modules.ModuleBuilder.Templates.IModSpec
             TemplateType = template.TemplateType();
             Role = template.GetRole();
             Location = template.GetDefaultLocation();
+            Severity = template.GetSeverity();
+            Classification = template.GetClassification();
         }
 
         public string SourceTemplateId { get; }
@@ -32,5 +36,7 @@ namespace Intent.Modules.ModuleBuilder.Templates.IModSpec
         public string TemplateType { get; set; }
         public string Role { get; }
         public string Location { get; }
+        public string Severity { get; }
+        public string Classification { get; }
     }
 }